+++ NEWS.md
@@ -1,6 +1,6 @@
 # ggalluvial 0.5.0.9001
 
-
+- `GeomFlow$draw_panel()` now begins by removing `complete.cases()`, corresponding to one-sided flows. (These are not removed in `StatFlow$compute_panel()`, which would exclude missing aesthetic values from legends.)
 
 # ggalluvial 0.5.0.9000